Centralize DataRow to PSObject conversion
