+++ DbaClientX.PowerShell/CmdletIInvokeDbaXQuery.cs
@@ -1,4 +1,3 @@
-﻿using System.Runtime.CompilerServices;
 using System.Data.Common;
 using System.Data.SqlClient;
 
@@ -166,7 +165,7 @@ public sealed class CmdletIInvokeDbaXQuery : AsyncPSCmdlet {
                     default:
                         await foreach (var row in enumerable.ConfigureAwait(false))
                         {
-                            WriteObject(DataRowToPSObject(row));
+                            WriteObject(PSObjectConverter.DataRowToPSObject(row));
                         }
                         break;
                 }
@@ -189,7 +188,7 @@ public sealed class CmdletIInvokeDbaXQuery : AsyncPSCmdlet {
                 if (ReturnType == ReturnType.PSObject) {
                     //var resultConverted = result as DataTable;
                     foreach (DataRow row in ((DataTable)result).Rows) {
-                        WriteObject(DataRowToPSObject(row));
+                        WriteObject(PSObjectConverter.DataRowToPSObject(row));
                     }
                 } else {
                     WriteObject(result, true);
@@ -203,35 +202,4 @@ public sealed class CmdletIInvokeDbaXQuery : AsyncPSCmdlet {
         }
     }
 
-    /// <summary>
-    /// Convert DataRow to PSObject
-    /// </summary>
-    /// <param name="row"></param>
-    /// <returns></returns>
-    private static readonly ConditionalWeakTable<DataTable, PSNoteProperty[]> _psNotePropertyCache = new();
-
-    private static PSObject DataRowToPSObject(DataRow row) {
-        PSObject psObject = new PSObject();
-
-        if (row != null && (row.RowState & DataRowState.Detached) != DataRowState.Detached) {
-            var table = row.Table;
-            if (!_psNotePropertyCache.TryGetValue(table, out var propertyTemplates)) {
-                propertyTemplates = new PSNoteProperty[table.Columns.Count];
-                for (int i = 0; i < table.Columns.Count; i++) {
-                    propertyTemplates[i] = new PSNoteProperty(table.Columns[i].ColumnName, null);
-                }
-                _psNotePropertyCache.Add(table, propertyTemplates);
-            }
-
-            for (int i = 0; i < propertyTemplates.Length; i++) {
-                var prop = (PSNoteProperty)propertyTemplates[i].Copy();
-                if (!row.IsNull(i)) {
-                    prop.Value = row[i];
-                }
-                psObject.Properties.Add(prop);
-            }
-        }
-
-        return psObject;
-    }
 }
