+++ DbaClientX.PowerShell/CmdletInvokeDbaXMySql.cs
@@ -1,4 +1,3 @@
-using System.Runtime.CompilerServices;
 
 namespace DBAClientX.PowerShell;
 
@@ -130,7 +129,7 @@ public sealed class CmdletInvokeDbaXMySql : AsyncPSCmdlet {
                         break;
                     default:
                         await foreach (var row in enumerable.ConfigureAwait(false)) {
-                            WriteObject(DataRowToPSObject(row));
+                            WriteObject(PSObjectConverter.DataRowToPSObject(row));
                         }
                         break;
                 }
@@ -145,7 +144,7 @@ public sealed class CmdletInvokeDbaXMySql : AsyncPSCmdlet {
             if (result != null) {
                 if (ReturnType == ReturnType.PSObject) {
                     foreach (DataRow row in ((DataTable)result).Rows) {
-                        WriteObject(DataRowToPSObject(row));
+                        WriteObject(PSObjectConverter.DataRowToPSObject(row));
                     }
                 } else {
                     WriteObject(result, true);
@@ -159,30 +158,4 @@ public sealed class CmdletInvokeDbaXMySql : AsyncPSCmdlet {
         }
     }
 
-    private static readonly ConditionalWeakTable<DataTable, PSNoteProperty[]> _psNotePropertyCache = new();
-
-    private static PSObject DataRowToPSObject(DataRow row) {
-        PSObject psObject = new PSObject();
-
-        if (row != null && (row.RowState & DataRowState.Detached) != DataRowState.Detached) {
-            var table = row.Table;
-            if (!_psNotePropertyCache.TryGetValue(table, out var propertyTemplates)) {
-                propertyTemplates = new PSNoteProperty[table.Columns.Count];
-                for (int i = 0; i < table.Columns.Count; i++) {
-                    propertyTemplates[i] = new PSNoteProperty(table.Columns[i].ColumnName, null);
-                }
-                _psNotePropertyCache.Add(table, propertyTemplates);
-            }
-
-            for (int i = 0; i < propertyTemplates.Length; i++) {
-                var prop = (PSNoteProperty)propertyTemplates[i].Copy();
-                if (!row.IsNull(i)) {
-                    prop.Value = row[i];
-                }
-                psObject.Properties.Add(prop);
-            }
-        }
-
-        return psObject;
-    }
 }
