+++ DbaClientX.Tests/PSObjectConverterTests.cs
@@ -1,10 +1,9 @@
 using System.Data;
 using System.Management.Automation;
-using System.Reflection;
 using DBAClientX.PowerShell;
 using Xunit;
 
-public class CmdletIInvokeDbaXQueryTests
+public class PSObjectConverterTests
 {
     [Fact]
     public void DataRowToPSObjectCreatesProperties()
@@ -21,14 +20,11 @@ public class CmdletIInvokeDbaXQueryTests
         row2["Name"] = "Bob";
         table.Rows.Add(row2);
 
-        var method = typeof(CmdletIInvokeDbaXQuery).GetMethod("DataRowToPSObject", BindingFlags.NonPublic | BindingFlags.Static);
-        Assert.NotNull(method);
-
-        var ps1 = (PSObject)method!.Invoke(null, new object[] { row1 })!;
+        var ps1 = PSObjectConverter.DataRowToPSObject(row1);
         Assert.Equal(1, ps1.Properties["Id"].Value);
         Assert.Equal("Alice", ps1.Properties["Name"].Value);
 
-        var ps2 = (PSObject)method.Invoke(null, new object[] { row2 })!;
+        var ps2 = PSObjectConverter.DataRowToPSObject(row2);
         Assert.Equal(2, ps2.Properties["Id"].Value);
         Assert.Equal("Bob", ps2.Properties["Name"].Value);
     }
